requirements.txt
plotly==4.7.1
pandas==1.3.1
wordcloud==1.7.0
gensim==4.0.1
newspaper3k
nltk==3.6.2
pyLDAvis==3.3.1
#scikit_learn==0.24.2
spacy==3.1.1
sumy==0.8.1
validators==0.18.2
